backend/requirements.txt
APScheduler==3.9.1
attrs==21.4.0
backports.zoneinfo==0.2.1
blinker==1.4
click==8.0.4
flasgger==0.9.5
Flask==2.0.3
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mistune==2.0.2
psycopg==3.0.11
psycopg-binary==3.0.11
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.0.3
zipp==3.7.0
